+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Occlusion test setups are now provided in a toml configuration file.
 - Added test executor and dumping of test results.
 - Added github action to run tests on push and pull requests.
+- Added naive ray casting for occlusion culling.
 
 ### Changed